requirements.txt
# Core dependencies
python-dotenv==1.0.0
streamlit==1.31.0

# Azure OpenAI
openai==1.12.0
azure-identity==1.15.0

# LLM Libraries
## Agno
agno==0.1.0

## LangChain
langchain==0.1.9
langchain-openai==0.0.8
langchain-community==0.0.24

## LangGraph
langgraph==0.0.26

## CrewAI
crewai==0.22.5
crewai-tools==0.1.7

## AutoGen
pyautogen==0.2.18

## LlamaIndex
llama-index==0.10.12
llama-index-llms-azure-openai==0.1.5
llama-index-embeddings-azure-openai==0.1.6

# Utilities
pydantic==2.6.1
tiktoken==0.6.0
numpy==1.24.3
pandas==2.1.4
asyncio==3.4.3
aiohttp==3.9.3
requests==2.31.0
typing-extensions==4.9.0

# Development
pytest==8.0.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
